framework-pkcs15.c: In function ‘set_gost_params’: framework-pkcs15.c:1892: warning: comparison between signed and unsigned framework-pkcs15.c:1902: warning: comparison between signed and unsigned
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@5529 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/pkcs11/framework-pkcs15.c
@@ -1884,7 +1884,7 @@ set_gost_params(struct sc_pkcs15init_keyarg_gost_params *first_params,
 	else
 		rv = attr_find(pPubTpl, ulPubCnt, CKA_GOSTR3410_PARAMS, &gost_params_oid, &len);
 	if (rv == CKR_OK) {
-		int nn = sizeof(gostr3410_param_oid)/sizeof(gostr3410_param_oid[0]);
+		size_t nn = sizeof(gostr3410_param_oid)/sizeof(gostr3410_param_oid[0]);
 
 		if (len != GOST_PARAMS_OID_SIZE)
 			return CKR_ATTRIBUTE_VALUE_INVALID;
